(broken) exchanged underlying graphics library of othuum project to ahwassa: it at least compiles

+++ Othuum/SelenNavigationLib/NavigationAgent.h
@@ -65,11 +65,11 @@ namespace Selen {
       _maxForce = speed;
     }
 
-    float getMaxSpeed() {
+    double getMaxSpeed() {
       return _maxSpeed;
     }
 
-    float getMaxForce() {
+    double getMaxForce() {
       return _maxForce;
     }
